+++ example/lib/main.dart
@@ -1,3 +1,5 @@
+import 'dart:typed_data';
+
 import 'package:ogg_caf_converter/ogg_caf_converter.dart';
 
 const String inputFilePath = 'path/to/input/file.opus'; // Input file location
@@ -22,6 +24,24 @@ void main() async {
       input: inputFilePath,
       output: outputFilePath,
       deleteInput: true,
+    );
+  } catch (e) {
+    // Handle error
+  }
+
+  // Convert from OGG to CAF in memory
+  try {
+    final Uint8List bytes = await OggCafConverter().convertOggToCafInMemory(
+      input: inputFilePath,
+    );
+  } catch (e) {
+    // Handle error
+  }
+
+  // Convert from CAF to OGG in memory
+  try {
+    final Uint8List bytes = await OggCafConverter().convertCafToOggInMemory(
+      input: inputFilePath,
     );
   } catch (e) {
     // Handle error